+++ dataworkspace/dataworkspace/apps/accounts/management/commands/store_db_creds_in_s3.py
@@ -2,7 +2,13 @@ from django.conf import settings
 from django.contrib.auth import get_user_model
 from django.core.management.base import BaseCommand
 
-from dataworkspace.apps.core.utils import new_private_database_credentials
+from dataworkspace.apps.core.utils import (
+    db_role_schema_suffix_for_user,
+    new_private_database_credentials,
+    postgres_user,
+    source_tables_for_user,
+    write_credentials_to_bucket,
+)
 
 
 class Command(BaseCommand):
@@ -17,7 +23,13 @@ class Command(BaseCommand):
         all_users = get_user_model().objects.order_by('last_name', 'first_name', 'id')
         for user in all_users:
             self.stdout.write(f'Creating credentials for {user.email}')
-            creds = new_private_database_credentials(user)
+
+            source_tables = source_tables_for_user(user)
+            db_role_schema_suffix = db_role_schema_suffix_for_user(user)
+            creds = new_private_database_credentials(
+                db_role_schema_suffix, source_tables, postgres_user(user.email)
+            )
+            write_credentials_to_bucket(user, creds)
             self.stdout.write(str(creds))
 
         self.stdout.write(self.style.SUCCESS('store_db_creds_in_s3 finished'))