+++ test/test_application.py
@@ -266,7 +266,7 @@ class TestApplication(unittest.TestCase):
 
         await asyncio.sleep(10)
 
-        stdout, stderr, code = await create_visualisation()
+        stdout, stderr, code = await create_visualisation_echo('testvisualisation')
         self.assertEqual(stdout, b'')
         self.assertEqual(stderr, b'')
         self.assertEqual(code, 0)
@@ -274,7 +274,7 @@ class TestApplication(unittest.TestCase):
         # Ensure the user doesn't see the visualisation link on the home page
         async with session.request('GET', 'http://localapps.com:8000/') as response:
             content = await response.text()
-        self.assertNotIn('Test Visualisation', content)
+        self.assertNotIn('Test testvisualisation', content)
 
         # Ensure the user doesn't have access to the application
         async with session.request(
@@ -285,7 +285,7 @@ class TestApplication(unittest.TestCase):
         self.assertIn('Forbidden', content)
         self.assertEqual(response.status, 403)
 
-        stdout, stderr, code = await give_user_visualisation_perms()
+        stdout, stderr, code = await give_user_visualisation_perms('testvisualisation')
         self.assertEqual(stdout, b'')
         self.assertEqual(stderr, b'')
         self.assertEqual(code, 0)
@@ -295,7 +295,7 @@ class TestApplication(unittest.TestCase):
         ) as response:
             application_content_1 = await response.text()
 
-        self.assertIn('Starting Test Visualisation', application_content_1)
+        self.assertIn('Starting Test testvisualisation', application_content_1)
 
         await asyncio.sleep(10)
 
@@ -314,6 +314,111 @@ class TestApplication(unittest.TestCase):
             received_content['headers']['from-downstream'], 'downstream-header-value'
         )
         self.assertEqual(received_headers['from-upstream'], 'upstream-header-value')
+
+    @async_test
+    async def test_visualisation_shows_dataset_if_authorised(self):
+        await flush_database()
+        await flush_redis()
+
+        session, cleanup_session = client_session()
+        self.add_async_cleanup(cleanup_session)
+
+        cleanup_application_1 = await create_application()
+        self.add_async_cleanup(cleanup_application_1)
+
+        is_logged_in = True
+        codes = iter(['some-code'])
+        tokens = iter(['token-1'])
+        auth_to_me = {
+            'Bearer token-1': {
+                'email': 'test@test.com',
+                'related_emails': [],
+                'first_name': 'Peter',
+                'last_name': 'Piper',
+                'user_id': '7f93c2c7-bc32-43f3-87dc-40d0b8fb2cd2',
+            }
+        }
+        sso_cleanup, _ = await create_sso(is_logged_in, codes, tokens, auth_to_me)
+        self.add_async_cleanup(sso_cleanup)
+
+        await asyncio.sleep(10)
+
+        # Ensure user created
+        async with session.request('GET', 'http://localapps.com:8000/') as response:
+            await response.text()
+
+        stdout, stderr, code = await create_private_dataset()
+        print(stderr)
+        self.assertEqual(stdout, b'')
+        self.assertEqual(stderr, b'')
+        self.assertEqual(code, 0)
+
+        stdout, stderr, code = await create_visualisation_dataset('testvisualisation-a')
+        self.assertEqual(stdout, b'')
+        self.assertEqual(stderr, b'')
+        self.assertEqual(code, 0)
+
+        stdout, stderr, code = await give_user_visualisation_perms(
+            'testvisualisation-a'
+        )
+        self.assertEqual(stdout, b'')
+        self.assertEqual(stderr, b'')
+        self.assertEqual(code, 0)
+
+        async with session.request(
+            'GET', 'http://testvisualisation-a.localapps.com:8000/'
+        ) as response:
+            application_content_1 = await response.text()
+
+        self.assertIn('Starting Test testvisualisation-a', application_content_1)
+
+        await asyncio.sleep(10)
+
+        async with session.request(
+            'GET', 'http://testvisualisation-a.localapps.com:8000/http'
+        ) as response:
+            received_status = response.status
+            await response.text()
+
+        # This application should not have permission to access the database
+        self.assertEqual(received_status, 500)
+
+        stdout, stderr, code = await create_visualisation_dataset('testvisualisation-b')
+        self.assertEqual(stdout, b'')
+        self.assertEqual(stderr, b'')
+        self.assertEqual(code, 0)
+
+        stdout, stderr, code = await give_user_visualisation_perms(
+            'testvisualisation-b'
+        )
+        self.assertEqual(stdout, b'')
+        self.assertEqual(stderr, b'')
+        self.assertEqual(code, 0)
+
+        stdout, stderr, code = await give_visualisation_dataset_perms(
+            'testvisualisation-b'
+        )
+        self.assertEqual(stdout, b'')
+        self.assertEqual(stderr, b'')
+        self.assertEqual(code, 0)
+
+        async with session.request(
+            'GET', 'http://testvisualisation-b.localapps.com:8000/'
+        ) as response:
+            application_content_2 = await response.text()
+
+        self.assertIn('Starting Test testvisualisation-b', application_content_2)
+
+        await asyncio.sleep(10)
+
+        async with session.request(
+            'GET', 'http://testvisualisation-b.localapps.com:8000/http'
+        ) as response:
+            received_status = response.status
+            received_content = await response.json()
+
+        self.assertEqual(received_content, {'data': [1]})
+        self.assertEqual(received_status, 200)
 
     @async_test
     async def test_application_custom_cpu_memory(self):
@@ -1366,9 +1471,9 @@ async def give_user_dataset_perms():
     return stdout, stderr, code
 
 
-async def give_user_visualisation_perms():
+async def give_user_visualisation_perms(name):
     python_code = textwrap.dedent(
-        """\
+        f"""\
         from django.contrib.auth.models import (
             User,
         )
@@ -1378,7 +1483,7 @@ async def give_user_visualisation_perms():
         )
         user = User.objects.get(profile__sso_id="7f93c2c7-bc32-43f3-87dc-40d0b8fb2cd2")
         visualisationtemplate = VisualisationTemplate.objects.get(
-            name="testvisualisation",
+            name="{name}",
         )
         ApplicationTemplateUserPermission.objects.create(
             application_template=visualisationtemplate,
@@ -1426,24 +1531,92 @@ async def make_table_google_data_studio_accessible():
     return stdout, stderr, code
 
 
-async def create_visualisation():
+async def create_visualisation_echo(name):
     python_code = textwrap.dedent(
-        """\
+        f"""\
         from dataworkspace.apps.applications.models import (
             VisualisationTemplate,
         )
         template = VisualisationTemplate.objects.create(
-            name="testvisualisation",
-            host_exact="testvisualisation",
-            host_pattern="testvisualisation",
-            nice_name="Test Visualisation",
+            name="{name}",
+            host_exact="{name}",
+            host_pattern="{name}",
+            nice_name="Test {name}",
             spawner="PROCESS",
-            spawner_options='{"CMD":["python3", "/test/echo_server.py"]}',
+            spawner_options='{{"CMD":["python3", "/test/echo_server.py"]}}',
             spawner_time=60,
             user_access_type="REQUIRES_AUTHORIZATION",
         )
         """
     ).encode('ascii')
+    give_perm = await asyncio.create_subprocess_shell(
+        'django-admin shell',
+        env=os.environ,
+        stdin=asyncio.subprocess.PIPE,
+        stdout=asyncio.subprocess.PIPE,
+        stderr=asyncio.subprocess.PIPE,
+    )
+    stdout, stderr = await give_perm.communicate(python_code)
+    code = await give_perm.wait()
+
+    return stdout, stderr, code
+
+
+async def create_visualisation_dataset(name):
+    python_code = textwrap.dedent(
+        f"""\
+        from dataworkspace.apps.applications.models import (
+            VisualisationTemplate,
+        )
+        template = VisualisationTemplate.objects.create(
+            name="{name}",
+            host_exact="{name}",
+            host_pattern="{name}",
+            nice_name="Test {name}",
+            spawner="PROCESS",
+            spawner_options='{{"CMD":["python3", "/test/dataset_server.py"]}}',
+            spawner_time=60,
+            user_access_type="REQUIRES_AUTHORIZATION",
+        )
+        """
+    ).encode('ascii')
+    give_perm = await asyncio.create_subprocess_shell(
+        'django-admin shell',
+        env=os.environ,
+        stdin=asyncio.subprocess.PIPE,
+        stdout=asyncio.subprocess.PIPE,
+        stderr=asyncio.subprocess.PIPE,
+    )
+    stdout, stderr = await give_perm.communicate(python_code)
+    code = await give_perm.wait()
+
+    return stdout, stderr, code
+
+
+async def give_visualisation_dataset_perms(name):
+    python_code = textwrap.dedent(
+        f"""\
+        from django.contrib.auth.models import (
+            User,
+        )
+        from dataworkspace.apps.applications.models import (
+            ApplicationTemplate,
+        )
+        from dataworkspace.apps.datasets.models import (
+            DataSet,
+            DataSetApplicationTemplatePermission,
+        )
+        application_template = ApplicationTemplate.objects.get(name="{name}")
+        dataset = DataSet.objects.get(
+            name="test_dataset",
+        )
+        DataSetApplicationTemplatePermission.objects.create(
+            application_template=application_template,
+            dataset=dataset,
+        )
+        """
+    ).encode('ascii')
+
     give_perm = await asyncio.create_subprocess_shell(
         'django-admin shell',
         env=os.environ,